Fix heap-use-after-free in serializeObject
serializeObject dereferenced a pointer with no guarantee that the pointed
data is still valid when the function is executed.

ServiceBoundObject::_gethost() should return a weak_ptr to let the
callbacks lock it and check if the host is still alive.

TSAN report that revealed the issue:
http://pastebin.aldebaran.lan/view/4cc5ff91
backtraces of a segfault that exposes the issue:
http://pastebin.aldebaran.lan/view/adc7f116

Note: serializeObject also takes a StreamContext* argument, there is no
guarantee that it is valid either during the callback execution.
I don't think we attempt to dereference it anywhere though, so I did not
change that.

Additionally:

 - added comments clarifying untrivial parts of the related code;
 - revealed and fixed invalid iterator accesses in Gateway impl;
 - removed unused signal;

Change-Id: I385c8b8ef27bdb30a05f101a10916bf234fe9065
Reviewed-on: http://gerrit.aldebaran.lan/76908
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>