Signal: fix disconnection deadlocks and get rid of "waitForInactive"
Considering the fact that onSubscribers can be given by anyone to a Signal,
it is extremely dangerous to call it synchronously from a critical section!
This caused deadlocks in separate projects, that could not be reproduced
in the attached tests. The attached tests were however still improving the
coverage and worth keeping.

As a consequence, the fix revealed the inconsistency of the
"waitForInactive" feature, that causes more problems it tries to solve:
it tries to bring a guarantee that callbacks are not running anymore
after the signal is disconnected, but it should be the user's
responsibility to manage the lifetime of his or her callbacks and use
mutexing, stranding and tracking to prevent doing bad things.

By trying to do too much, we ended up waiting in very complex
configurations, causing deadlocks that the user could not even predict.

Change-Id: I3b4a01f5c8441f6f3963769b8b2f2c944142ffc6
Reviewed-on: http://gerrit.aldebaran.lan/80711
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: vpaleologue <vpaleologue@aldebaran-robotics.com>